+++ src/app/(route)/customer/(route)/feedback/feedback-info/[feedbackId]/_constants/InfoItems.ts
@@ -1,4 +1,6 @@
-export const InfoItems = (feedbackInfoData: any) => [  // TODO: 타입 변경
+import { FeedbackInfoType } from "../../../_types/FeedbackInfoType";
+
+export const InfoItems = (feedbackInfoData: FeedbackInfoType) => [
   { label: "조회수", value: feedbackInfoData?.viewCount },
   { label: "댓글", value: feedbackInfoData?.commentCount },
   { label: "추천", value: feedbackInfoData?.recommendCount },
